[JENKINS-45273] Extract duplicated code into its own class
